feat: update qemu to 1%5.2+dfsg
fast processor emulator, dummy package

Issue: https://github.com/linuxdeepin/developer-center/issues/7284
Log: update repo

+++ tests/qtest/tpm-tests.c
@@ -1,5 +1,5 @@
 /*
- * QTest TPM common test code
+ * QTest TPM commont test code
  *
  * Copyright (c) 2018 IBM Corporation
  * Copyright (c) 2018 Red Hat, Inc.
@@ -123,10 +123,14 @@ void tpm_test_swtpm_migration_test(const char *src_tpm_path,
     qtest_quit(src_qemu);
 
     tpm_util_swtpm_kill(dst_tpm_pid);
-    g_unlink(dst_tpm_addr->u.q_unix.path);
-    qapi_free_SocketAddress(dst_tpm_addr);
+    if (dst_tpm_addr) {
+        g_unlink(dst_tpm_addr->u.q_unix.path);
+        qapi_free_SocketAddress(dst_tpm_addr);
+    }
 
     tpm_util_swtpm_kill(src_tpm_pid);
-    g_unlink(src_tpm_addr->u.q_unix.path);
-    qapi_free_SocketAddress(src_tpm_addr);
+    if (src_tpm_addr) {
+        g_unlink(src_tpm_addr->u.q_unix.path);
+        qapi_free_SocketAddress(src_tpm_addr);
+    }
 }
